+++ packages/graphql-resolver-utils/src/index.ts
@@ -1,2 +1,3 @@
 export { entityResolver } from "./entityResolver";
+export { convertInfoToLoadHint } from "./hint";
 export { saveEntities, saveEntity } from "./saveEntity";