+++ packages/orm/src/relations/ManyToOneReference.ts
@@ -15,6 +15,7 @@ import {
   setField,
   tagId,
 } from "../index";
+import { maybeAdd, maybeRemove } from "../utils";
 import { AbstractRelationImpl } from "./AbstractRelationImpl";
 import { OneToManyCollection } from "./OneToManyCollection";
 import { ReferenceN } from "./Reference";
@@ -37,7 +38,6 @@ export function isManyToOneReference(maybeReference: any): maybeReference is Man
 
 export interface ManyToOneReference<T extends Entity, U extends Entity, N extends never | undefined>
   extends Reference<T, U, N> {
-
   /** Returns the id of the current assigned entity, or a runtime error if either 1) unset or 2) set to a new entity that doesn't have an `id` yet. */
   id: IdOf<U>;
 
@@ -116,6 +116,15 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
     return this._isLoaded;
   }
 
+  get isPreloaded(): boolean {
+    return !!this.maybeFindEntity();
+  }
+
+  preload(): void {
+    this.loaded = this.maybeFindEntity();
+    this._isLoaded = true;
+  }
+
   private doGet(opts?: { withDeleted?: boolean }): U | N {
     ensureNotDeleted(this.#entity, "pending");
     // This should only be callable in the type system if we've already resolved this to an instance,
@@ -152,6 +161,7 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
       id = tagId(this.otherMeta, id) as IdOf<U>;
     }
 
+    const previousId = this.idTagged;
     const previous = this.maybeFindEntity();
     const changed = setField(this.#entity, this.fieldName, id);
     if (!changed) {
@@ -160,7 +170,7 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
 
     this.loaded = id ? this.#entity.em.getEntity(id) : undefined;
     this._isLoaded = !!this.loaded;
-    this.maybeRemove(previous);
+    this.maybeRemove(previousId, previous);
     this.maybeAdd();
   }
 
@@ -208,6 +218,7 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
       return;
     }
 
+    const previousId = this.idTagged;
     const previous = this.maybeFindEntity();
     // Prefer to keep the id in our data hash, but if this is a new entity w/o an id, use the entity itself
     setField(this.#entity, this.fieldName, isEntity(other) ? other?.idTaggedMaybe ?? other : other);
@@ -219,7 +230,7 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
       this.loaded = other;
       this._isLoaded = true;
     }
-    this.maybeRemove(previous);
+    this.maybeRemove(previousId, previous);
     this.maybeAdd();
   }
 
@@ -267,7 +278,7 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
     this._isLoaded = true;
   }
 
-  maybeRemove(other: U | undefined) {
+  maybeRemove(otherId: string | undefined, other: U | undefined) {
     if (other) {
       const prevRelation = this.getOtherRelation(other);
       if (prevRelation instanceof OneToManyCollection) {
@@ -277,6 +288,21 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
       } else {
         prevRelation.set(undefined as any, { percolating: true });
       }
+    } else if (otherId) {
+      // Other is not loaded in memory, but cache it in case our other side is later loaded
+      const { em } = this.#entity;
+      let map = getEmInternalApi(em).pendingChildren.get(otherId);
+      if (!map) {
+        map = new Map();
+        getEmInternalApi(em).pendingChildren.set(otherId, map);
+      }
+      let pending = map.get(this.otherFieldName);
+      if (!pending) {
+        pending = { adds: [], removes: [] };
+        map.set(this.otherFieldName, pending);
+      }
+      maybeAdd(pending.removes, this.#entity);
+      maybeRemove(pending.adds, this.#entity);
     }
   }
 
@@ -301,12 +327,13 @@ export class ManyToOneReferenceImpl<T extends Entity, U extends Entity, N extend
         map = new Map();
         getEmInternalApi(em).pendingChildren.set(id, map);
       }
-      let list = map.get(this.otherFieldName);
-      if (!list) {
-        list = [];
-        map.set(this.otherFieldName, list);
+      let pending = map.get(this.otherFieldName);
+      if (!pending) {
+        pending = { adds: [], removes: [] };
+        map.set(this.otherFieldName, pending);
       }
-      list.push(this.#entity);
+      maybeAdd(pending.adds, this.#entity);
+      maybeRemove(pending.removes, this.#entity);
     }
   }