+++ packages/orm/src/relations/OneToOneReference.ts
@@ -1,4 +1,12 @@
-import { currentlyInstantiatingEntity, deTagId, ensureNotDeleted, IdOf, LoadedReference, setField } from "../";
+import {
+  currentlyInstantiatingEntity,
+  deTagId,
+  ensureNotDeleted,
+  getEmInternalApi,
+  IdOf,
+  LoadedReference,
+  setField,
+} from "../";
 import { oneToOneDataLoader } from "../dataloaders/oneToOneDataLoader";
 import { Entity } from "../Entity";
 import { EntityMetadata, getMetadata } from "../EntityMetadata";
@@ -146,7 +154,10 @@ export class OneToOneReferenceImpl<T extends Entity, U extends Entity>
     ensureNotDeleted(this.#entity, "pending");
     if (!this._isLoaded || opts.forceReload) {
       if (!this.#entity.isNewEntity) {
-        this.loaded = await oneToOneDataLoader(this.#entity.em, this).load(this.#entity.idTagged);
+        const joinLoaded = this.getPreloaded();
+        this.loaded = joinLoaded
+          ? joinLoaded[0]
+          : await oneToOneDataLoader(this.#entity.em, this).load(this.#entity.idTagged);
       }
       this._isLoaded = true;
     }
@@ -175,6 +186,15 @@ export class OneToOneReferenceImpl<T extends Entity, U extends Entity>
     return this._isLoaded;
   }
 
+  get isPreloaded(): boolean {
+    return !!this.getPreloaded();
+  }
+
+  preload(): void {
+    this.loaded = this.getPreloaded()?.[0];
+    this._isLoaded = true;
+  }
+
   get getWithDeleted(): U | undefined {
     return this.filterDeleted(this.doGet(), { withDeleted: true });
   }
@@ -195,7 +215,7 @@ export class OneToOneReferenceImpl<T extends Entity, U extends Entity>
     ensureNotDeleted(this.#entity, "pending");
     if (!this._isLoaded) {
       // This should only be callable in the type system if we've already resolved this to an instance
-      throw new Error("get was called when not preloaded");
+      throw new Error("get was called when not loaded");
     }
     return this.loaded;
   }
@@ -241,6 +261,11 @@ export class OneToOneReferenceImpl<T extends Entity, U extends Entity>
   /** Returns the other relation that points back at us, i.e. we're `Author.image` and this is `Image.author_id`. */
   private getOtherRelation(other: U): ManyToOneReference<U, T, any> {
     return (other as U)[this.otherFieldName] as any;
+  }
+
+  private getPreloaded(): U[] | undefined {
+    if (this.#entity.isNewEntity) return undefined;
+    return getEmInternalApi(this.#entity.em).getPreloadedRelation(this.#entity.idTagged, this.fieldName);
   }
 
   [RelationT] = null!;