+++ packages/plugins/join-preloading/jest.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+  transform: { "^.+\\.tsx?$": "@swc/jest" },
+  moduleNameMapper: {
+    "^@src/(.*)": "<rootDir>/src/$1",
+  },
+  // globalSetup: "<rootDir>/src/setupTestEnv.js",
+  testMatch: ["<rootDir>/src/**/*.test.(ts|tsx)"],
+  testEnvironment: "node",
+  reporters: [
+    "default",
+    [
+      "jest-junit",
+      {
+        outputDirectory: "../../../artifacts",
+        outputName: "junit-plugins-join-preloading.xml",
+        usePathForSuiteName: "true",
+      },
+    ],
+  ],
+};